+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Version 5
+
+- Replaced `IsoDate::equals(left, right)` with `IsoDate.equals(other)`
+
 # Version 4
 
 - Removed `getNowTimestamp()` - just use `Date.now()` instead